+++ protobuf/src/well_known_types/source_context.rs
@@ -89,13 +89,13 @@ impl crate::Message for SourceContext {
 
     // Compute sizes of nested messages
     #[allow(unused_variables)]
-    fn compute_size(&self) -> u32 {
+    fn compute_size(&self) -> u64 {
         let mut my_size = 0;
         if !self.file_name.is_empty() {
             my_size += crate::rt::string_size(1, &self.file_name);
         }
         my_size += crate::rt::unknown_fields_size(self.unknown_fields());
-        self.cached_size.set(my_size);
+        self.cached_size.set(my_size as u32);
         my_size
     }
